Add SyncBackground component
closes #904
